text_reader/requirements.txt
grpcio
protobuf
requests